Add Login and Registration with Email  Verification

+++ app/Http/Controllers/NoteController.php
@@ -12,7 +12,10 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::orderBy("created_at", "desc")->paginate(15);
+        $notes = Note::query()
+            ->where("user_id", auth()->user()->id)
+            ->orderBy("created_at", "desc")
+            ->paginate(15);
         // dd($notes);
         return view('note.index', ['notes' => $notes]);
     }
@@ -34,7 +37,7 @@ class NoteController extends Controller
             'note' => ['required', 'string'],
         ]);
 
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
         $note = Note::create($data);
 
         return redirect()->route('note.show', $note)->with('message', 'Note was created');
@@ -46,6 +49,10 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        if (auth()->user()->id != $note->user_id) {
+            abort(403);
+        }
+
         return view('note.show', ['note' => $note]);
     }
 
@@ -54,6 +61,10 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
+        if (auth()->user()->id != $note->user_id) {
+            abort(403);
+        }
+
         return view('note.edit', ['note' => $note]);
     }
 
@@ -62,6 +73,10 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        if (auth()->user()->id != $note->user_id) {
+            abort(403);
+        }
+
         $data = $request->validate([
             'note' => ['required', 'string'],
         ]);
@@ -76,6 +91,10 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if (auth()->user()->id != $note->user_id) {
+            abort(403);
+        }
+
         $note->delete();
 
         return redirect()->route('note.index')->with('message', 'Note was deleted');
